+++ tests/it/main.rs
@@ -2,6 +2,7 @@ mod basic;
 mod errors;
 mod generic;
 mod matching_eq;
+mod matching_pat;
 mod mixed;
 mod mock_order;
 mod prefix;