+++ unimock_macros/src/matching/mod.rs
@@ -240,7 +240,6 @@ impl ArgMatcher {
             ArgMatcher::Pattern(pat) => match &pat {
                 syn::Pat::Wild(_) => None,
                 pat => {
-                    let arg_ident = &arg.arg_ident;
                     let mut doc_string = String::new();
                     pat.doc(&mut doc_string);
 
@@ -250,9 +249,9 @@ impl ArgMatcher {
                     Some(quote! {
                         match #arg_expr {
                             #pat => {}
-                            _ => {
+                            mismatch => {
                                 use ::unimock::macro_api::{ProperDebug, NoDebug};
-                                reporter.pat_fail(#index, #arg_ident.unimock_try_debug(), #doc_lit);
+                                reporter.pat_fail(#index, mismatch.unimock_try_debug(), #doc_lit);
                             }
                         }
                     })