+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Documentation note about `#[cfg_attr(test, unimock)]`.
+### Fixed
+- Compile error when `matching!` a literal string pattern against a type that implements both `AsRef<str>` and `Debug`.
 
 ## [0.4.1] - 2022-11-24
 ### Added